chore: rename helpers.ts to api.ts for better clarity
