+++ src/index.ts
@@ -1,4 +1,4 @@
 export { miniFetch } from './core/miniFetch'
 export { createMiniFetch } from './core/factories'
-export { miniFetchApi } from './core/helpers'
+export { miniFetchApi } from './core/api'
 export { HttpError, TimeoutError, RequestError } from './errors'
